+++ src/components/RightPanel/index.vue
@@ -2,7 +2,10 @@
   <div ref="rightPanel" :class="{show:show}" class="rightPanel-container">
     <div class="rightPanel-background" />
     <div class="rightPanel">
-      <div :style="{'top':buttonTop+'px'}" @click="show=!show"/>
+      <div v-if="settingBtn" :style="{'top':buttonTop+'px','background-color':theme}" class="handle-button" @click="show=!show">
+        <i :class="show?'el-icon-close':'el-icon-setting'" />
+      </div>
+      <div v-else :style="{'top':buttonTop+'px'}" @click="show=!show"/>
       <div class="rightPanel-items">
         <slot />
       </div>
@@ -25,11 +28,6 @@ export default {
       type: Number
     }
   },
-  data() {
-    return {
-      theme: '#fff'
-    }
-  },
   computed: {
     show: {
       get() {
@@ -40,6 +38,16 @@ export default {
           key: 'showRightPanel',
           value: val
         })
+      }
+    },
+    theme: {
+      get() {
+        return this.$store.state.settings.theme
+      }
+    },
+    settingBtn: {
+      get() {
+        return this.$store.state.settings.settingBtn
       }
     }
   },
