+++ src/router/index.js
@@ -1,83 +1,68 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+import router from './routers'
+import store from '@/store'
+import Config from '@/config'
+import NProgress from 'nprogress' // progress bar
+import 'nprogress/nprogress.css'// progress bar style
+import { getToken } from '@/utils/auth' // getToken from cookie
+import { buildMenus } from '@/api/menu'
+import { filterAsyncRouter } from '@/store/modules/permission'
 
-Vue.use(Router)
+NProgress.configure({ showSpinner: false })// NProgress Configuration
 
-/* Layout */
-import Layout from '../views/layout/Layout'
+const whiteList = ['/login']// no redirect whitelist
 
-/**
-* hidden: true                   if `hidden:true` will not show in the sidebar(default is false)
-* alwaysShow: true               if set true, will always show the root menu, whatever its child routes length
-*                                if not set alwaysShow, only more than one route under the children
-*                                it will becomes nested mode, otherwise not show the root menu
-* redirect: noredirect           if `redirect:noredirect` will no redirect in the breadcrumb
-* name:'router-name'             the name is used by <keep-alive> (must set!!!)
-* meta : {
-    title: 'title'               the name show in submenu and breadcrumb (recommend set)
-    icon: 'svg-name'             the icon show in the sidebar,
+router.beforeEach((to, from, next) => {
+  if (to.meta.title) {
+    document.title = to.meta.title + ' - ' + Config.webName
   }
-**/
-
-export const constantRouterMap = [
-  { path: '/login',
-    meta: { title: '登录', noCache: true },
-    component: () => import('@/views/login/index'),
-    hidden: true
-  },
-  {
-    path: '/404',
-    component: () => import('@/views/errorPage/404'),
-    hidden: true
-  },
-  {
-    path: '/401',
-    component: () => import('@/views/errorPage/401'),
-    hidden: true
-  },
-  {
-    path: '/redirect',
-    component: Layout,
-    hidden: true,
-    children: [
-      {
-        path: '/redirect/:path*',
-        component: () => import('@/views/redirect/index')
+  NProgress.start() // start progress bar
+  if (getToken()) {
+    // 已登录且要跳转的页面是登录页
+    if (to.path === '/login') {
+      next({ path: '/' })
+      NProgress.done() // if current page is dashboard will not trigger	afterEach hook, so manually handle it
+    } else {
+      if (store.getters.roles.length === 0) { // 判断当前用户是否已拉取完user_info信息
+        store.dispatch('GetInfo').then(res => { // 拉取user_info
+          // 动态路由，拉取菜单
+          loadMenus(next, to)
+        }).catch((err) => {
+          console.log(err)
+          store.dispatch('LogOut').then(() => {
+            location.reload() // 为了重新实例化vue-router对象 避免bug
+          })
+        })
+      // 登录时未拉取 菜单，在此处拉取
+      } else if (store.getters.loadMenus) {
+        // 修改成false，防止死循环
+        store.dispatch('updateLoadMenus').then(res => {})
+        loadMenus(next, to)
+      } else {
+        next()
       }
-    ]
-  },
-  {
-    path: '/',
-    component: Layout,
-    redirect: 'dashboard',
-    children: [
-      {
-        path: 'dashboard',
-        component: () => import('@/views/dashboard/index'),
-        name: '首页',
-        meta: { title: '首页', icon: 'index', noCache: true, affix: true }
-      }
-    ]
-  },
-  {
-    path: '/user',
-    component: Layout,
-    hidden: true,
-    redirect: 'noredirect',
-    children: [
-      {
-        path: 'center',
-        component: () => import('@/views/system/user/center'),
-        name: '个人中心',
-        meta: { title: '个人中心', icon: 'user' }
-      }
-    ]
+    }
+  } else {
+    /* has no token*/
+    if (whiteList.indexOf(to.path) !== -1) { // 在免登录白名单，直接进入
+      next()
+    } else {
+      next(`/login?redirect=${to.path}`) // 否则全部重定向到登录页
+      NProgress.done()
+    }
   }
-  // { path: '*', redirect: '/404', hidden: true }
-]
+})
 
-export default new Router({
-  mode: 'history',
-  scrollBehavior: () => ({ y: 0 }),
-  routes: constantRouterMap
+export const loadMenus = (next, to) => {
+  buildMenus().then(res => {
+    const asyncRouter = filterAsyncRouter(res)
+    asyncRouter.push({ path: '*', redirect: '/404', hidden: true })
+    store.dispatch('GenerateRoutes', asyncRouter).then(() => { // 存储路由
+      router.addRoutes(asyncRouter) // 动态添加可访问路由表
+      next({ ...to, replace: true })// hack方法 确保addRoutes已完成
+    })
+  })
+}
+
+router.afterEach(() => {
+  NProgress.done() // finish progress bar
 })
