+++ src/views/login/index.vue
@@ -32,8 +32,8 @@ export default {
   data() {
     return {
       loginForm: {
-        username: '',
-        password: '',
+        username: 'admin',
+        password: '123456',
         rememberMe: false
       },
       loginRules: {
