+++ src/views/system/role/module/form.vue
@@ -137,5 +137,4 @@ export default {
 </script>
 
 <style scoped>
-
 </style>
